requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
class-registry==2.1.2
click==7.1.2
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.1.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
filters==1.3.2
idna==3.10
pillow==11.0.0
psycopg2==2.9.10
pycryptodome==3.21.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==8.0
regex==2024.9.11
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typer==0.3.2
typing_extensions==4.12.2
urllib3==2.2.3
